Misc updates. (#571)
Fix #530 
Fix #447 

Update rAthena cb2c553, 6d199ca, ac46920, 02c69f6, fc7e46e, bf52c52, 3caf668, 17f32a1, 3dd388a, 1658067, be1194e, 9b11301, 8431410, 152bdb0, a83e9f3, bc15f5c, beb6b57, 4664e88, 63c4c9c, c1112c2, ddb9e81, 3126bd8, ce42168, 0d1eb99, 62fa1b8, 752d314, 81c5dc9, 08d160a, f290bc9, df6385f

Update Hercules a3b6618, cde8cf6, 3cde523, 8842259, 238cfaa, a3bc9ad, 2e4a0bf

+++ src/common/grfio.c
@@ -407,7 +407,7 @@ void *grfio_reads(const char *fname, int *size)
 			fclose(in);
 
 			if( size )
-				*size = declen;
+				*size = (int)declen;
 		} else {
 			if (entry != NULL && entry->gentry < 0) {
 				entry->gentry = -entry->gentry;	// local file checked
